Change visualisation to use swing/awt
I decided to use swing/awt and use labels, panels and buutons
for the visualisation instead of drawing the lines and printing text
on the window by coordinates. When updating the visualisation
the window now doesn't flicker and it doesn't have to redraw out the
whole thing when updating values.

The actual visualisation now looks less sudoku-like, however this
can be sorted in the future.
[Ticket: X]