Implement recursive solving and csp
The constrains methods have been created to check is a particular
value for a square is allowed based on the sudoku constrains.

Additionally the sudoku can now be solved using recursion. On each
recusrive call the possible values for each square are calculated using
the constraint rules. Then the squares are then looped through and if a
value is 0 (ie not set) then the possible values are looped through.
The value is set and then a recursive call happens. If that call is
true then the value can be set if fine, otherwise the value is set
back to 0. The method itself will return true only if none of the
square values are 0, and hence the sudoku has been solved. If there is
a point where there is a square with a value of 0 and it has no
possible values, then that path taken dies not work and so the call
returns false, resulting in backtracking.