Fix solve and add timer
After testing the program on a few more complex puzzles, I found out
there was an error when solving in that the program would not
backtrack correctly. As such I have changed the way solving works
in order to make sure that all puzzles can be solved correctly
and backtracking works as intended.

Additionally I have added a sleep timer so that it is clear to see what
is going on when the program is running.

[Ticket: X]